# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Football AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface League {
  id: number;
  name: string;
  country: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  season?: number | null;
}

export interface Prediction {
  home_win_prob: number;
  draw_prob: number;
  away_win_prob: number;
  btts_prob: number;
  over_25_prob: number;
  predicted_winner: string;
  confidence: number;
  home_xg: number;
  away_xg: number;
  predicted_score: string;
}

export interface MatchWithPrediction {
  fixture_id: number;
  league_id: number;
  league_name: string;
  /** @nullable */
  league_logo?: string | null;
  /** @nullable */
  league_country?: string | null;
  home_team: string;
  /** @nullable */
  home_logo?: string | null;
  away_team: string;
  /** @nullable */
  away_logo?: string | null;
  kickoff: string;
  status: string;
  /** @nullable */
  home_score?: number | null;
  /** @nullable */
  away_score?: number | null;
  prediction: Prediction;
}

export type GetFootballMatchesTodayParams = {
/**
 * @nullable
 */
league_id?: number | null;
};

export type GetFootballPredictionsTodayParams = {
/**
 * @nullable
 */
league_id?: number | null;
};