# lib/api-zod/src/generated/types/league.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Football AI Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface League {
  id: number;
  name: string;
  country: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  season?: number | null;
}