# lib/api-zod/src/generated/types/matchWithPrediction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Football AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
import type { Prediction } from './prediction';

export interface MatchWithPrediction {
  fixture_id: number;
  league_id: number;
  league_name: string;
  /** @nullable */
  league_logo?: string | null;
  /** @nullable */
  league_country?: string | null;
  home_team: string;
  /** @nullable */
  home_logo?: string | null;
  away_team: string;
  /** @nullable */
  away_logo?: string | null;
  kickoff: string;
  status: string;
  /** @nullable */
  home_score?: number | null;
  /** @nullable */
  away_score?: number | null;
  prediction: Prediction;
}