# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Football AI Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  home_win_prob: number;
  draw_prob: number;
  away_win_prob: number;
  btts_prob: number;
  over_25_prob: number;
  predicted_winner: string;
  confidence: number;
  home_xg: number;
  away_xg: number;
  predicted_score: string;
}